requirements.txt
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.3.0
caches==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
datatypes==0.9.3
decorators==2.0.7
dnspython==2.3.0
dsnparse==0.2.0
fake-useragent==1.1.3
fastapi==0.95.1
fastapi-cache2==0.2.1
fastapi-pagination==0.12.3
frozenlist==1.3.3
h11==0.14.0
hiredis==2.2.3
httptools==0.5.0
idna==3.4
isort==5.12.0
kafka-python==2.0.2
lxml==4.9.2
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
pendulum==2.1.2
platformdirs==3.5.0
pydantic==1.10.7
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytzdata==2020.1
PyYAML==6.0
redis==4.5.5
requests==2.30.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
yarl==1.9.2
